refactor(core): `buildTreeEditingState`
