+++ packages/sanity/src/core/form/studio/tree-editing/utils/build-tree-editing-state/index.ts
@@ -1,0 +1,1 @@
+export * from './buildTreeEditingState'
