+++ packages/sanity/src/core/form/studio/tree-editing/utils/index.ts
@@ -1,3 +1,3 @@
-export * from './buildTreeEditingState'
+export * from './build-tree-editing-state'
 export * from './getSchemaField'
 export * from './shouldArrayDialogOpen'
